+++ slamrs/neato/src/connection.rs
